fix importing new tags
We were using the "response" of the image creation called as image name.
This was working when there was no previous image, but when that is the
case, response contains a message error.

This commit fixes it

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>
(cherry picked from commit a7844f96d2f19517700204560d474eb20cde9b5e)