add crio support
container-feeder defaults to loading the images into the Docker daemon,
which can be changed by setting the "feeder-target" field in the
/etc/container-feeder.json config to "crio".  The loading of the images
to the graphroot of crio is achieved by using the libpod library, adding
more dependencies to the vendor directory.

Notice, that this change implies some heavy refactorings.  Now, we have
a more generalized design and a type hiearchry that will ultimately ease
maintenance.  For packaging purposes a default container-feeder.json
file is added to the root of the project.

Further changes:
  * Add --log-level option to facilitate debugging.
  * Manually update (and fix) containers/image and libpod to load xz
    compressed files. Notice that once libpod has fixed the issue, we
    need to update our ./vendor directory.
  * Update Travis:
    * Use go 1.8.x and 1.9.x
    * Perform fmt checks via govendor
    * Replace go test by a more meaningful compile-test

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>