add whitelist support
Depending on the cluster-role and use-case of a machine, only a subset
of container images must be loaded.  Add whitelist support for image
loading to container-feeder in order to avoid redundant loading and to
ultimately speed things up.

The whitelist can be specified via the `/etc/container-feeder.json`
config.  The `whitelist` field in the config is an array of strings,
each representing an image name (e.g., `opensuse/tumbleweed`).  Hence,
the whitelist filtering is image-name based and ignores tags.  If the
whitelist is empty, no filtering will be applied.

From the code perspective, the change required to break the Feeder into
a concrete type and an interface.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>
feature#crio